Refactor tools and annotations.
Tools are now managed by an instance of ToolManager, which is owned by PhotoFrame (and not its view!).

The annotations must implement the Annotation<S> interface, where S is the CShape they are (this is required to display them in the annotation canvas). They are part of a PhotoFrame in two ways: (1) as annotations in the model and (2) as shapes in the canvas of the view.

Duplicating the shape type as a type parameter of Annotation was necessary because of Java's type system limitations/the absence of interface for shapes in the SwingStates library.

Various parts of the code had to be slightly refactored to handle the new responsabilities of the model/view/controller, as well as to fix type issues related to what is explained above.